Move torchair_attention to torchair dir
